requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==24.1
django-rename-app==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
